+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.1] - 2023-06-21
 - improve performance AVR
 - add percentage wrappers
-- add channel guard in setPWM().
+- add channel guard in **setPWM()**
+- changed return type **setPWM()** to bool.
+- update keywords.txt
 - minor edits.